+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Check results now show useful info and error messages
 - The venv created by pygrader is now with a non-standard name
 - Deleting existing venv is now optional
+- If a zip contains only the project as a subdirectory, use the subdir instead
 
 ## 1.7.1